# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Conveer API
  description: |
    API для управления автоматизированным конвейером создания и прогрева аккаунтов 
    в социальных сетях (VK, Telegram, Mail.ru, Max).
    
    ## Аутентификация
    API использует Bearer токен для аутентификации. Токен можно получить через Telegram бота.
    
    ## Rate Limiting
    - 100 запросов в минуту для обычных endpoint'ов
    - 10 запросов в минуту для операций создания аккаунтов
    
    ## Коды ошибок
    - `400` - Некорректный запрос
    - `401` - Не авторизован
    - `403` - Доступ запрещён
    - `404` - Ресурс не найден
    - `429` - Превышен лимит запросов
    - `500` - Внутренняя ошибка сервера
  version: 1.0.0
  contact:
    name: Conveer Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api/v1
    description: Development server
  - url: https://api.conveer.example.com/v1
    description: Production server
...
paths:
...
  /warming/tasks:
    get:
      tags:
        - warming
      summary: Получить список задач прогрева
      operationId: listWarmingTasks
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
        - name: status
          in: query
          schema:
            $ref: '#/components/schemas/WarmingStatus'
        - name: account_id
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Список задач прогрева
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WarmingTask'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - warming
      summary: Запустить прогрев аккаунта
      operationId: startWarming
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartWarmingRequest'
      responses:
        '201':
          description: Задача прогрева создана
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmingTask'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Прогрев уже запущен для данного аккаунта
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /warming/tasks/{taskId}/pause:
    post:
      tags:
        - warming
      summary: Приостановить прогрев
      operationId: pauseWarming
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Прогрев приостановлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmingTask'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /warming/tasks/{taskId}/resume:
    post:
      tags:
        - warming
      summary: Возобновить прогрев
      operationId: resumeWarming
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Прогрев возобновлён
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmingTask'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /warming/tasks/{taskId}/stop:
    post:
      tags:
        - warming
      summary: Остановить прогрев
      operationId: stopWarming
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Прогрев остановлен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmingTask'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /warming/tasks/{taskId}/statistics:
    get:
      tags:
        - warming
      summary: Получить статистику прогрева
      operationId: getWarmingStatistics
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TaskIdParam'
      responses:
        '200':
          description: Статистика прогрева
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmingStatistics'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /warming/scenarios:
    get:
      tags:
        - warming
      summary: Получить список сценариев прогрева
      operationId: listScenarios
      security:
        - bearerAuth: []
      parameters:
        - name: platform
          in: query
          schema:
            $ref: '#/components/schemas/Platform'
      responses:
        '200':
          description: Список сценариев
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WarmingScenario'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - warming
      summary: Создать кастомный сценарий
      operationId: createScenario
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateScenarioRequest'
      responses:
        '201':
          description: Сценарий создан
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarmingScenario'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ==================== SMS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный через Telegram бота
  parameters:
    PageParam:
      name: page
      in: query
      description: Номер страницы
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Количество элементов на странице
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    TaskIdParam:
      name: taskId
      in: path
      required: true
      description: ID задачи прогрева
      schema:
        type: string
        pattern: '^[a-f0-9]{24}$'
  responses:
    BadRequest:
      description: Некорректный запрос
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Не авторизован
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "UNAUTHORIZED"
            message: "Invalid or expired token"
    NotFound:
      description: Ресурс не найден
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Код ошибки
        message:
          type: string
          description: Описание ошибки
        details:
          type: object
          additionalProperties: true
          description: Дополнительные детали ошибки
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
...
    Platform:
      type: string
      enum: [vk, telegram, mail, max]
      description: Поддерживаемые платформы

    # ==================== Account Schemas ====================
...
    WarmingStatus:
      type: string
      enum:
        - pending
        - running
        - paused
        - completed
        - failed
      description: Статус задачи прогрева
    WarmingTask:
      type: object
      properties:
        id:
          type: string
        account_id:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        scenario_id:
          type: string
        status:
          $ref: '#/components/schemas/WarmingStatus'
        progress:
          type: integer
          minimum: 0
          maximum: 100
        current_day:
          type: integer
        total_days:
          type: integer
        actions_completed:
          type: integer
        actions_failed:
          type: integer
        next_action_at:
          type: string
          format: date-time
          nullable: true
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
          nullable: true
        error:
          type: string
          nullable: true
    WarmingStatistics:
      type: object
      properties:
        task_id:
          type: string
        total_actions:
          type: integer
        successful_actions:
          type: integer
        failed_actions:
          type: integer
        actions_by_type:
          type: object
          additionalProperties:
            type: integer
        errors_by_type:
          type: object
          additionalProperties:
            type: integer
        average_delay_seconds:
          type: number
        total_duration_seconds:
          type: integer
    WarmingScenario:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        platform:
          $ref: '#/components/schemas/Platform'
        type:
          type: string
          enum: [basic, advanced, custom]
        duration_days:
          type: integer
        daily_actions:
          type: array
          items:
            $ref: '#/components/schemas/DailyActionPlan'
        is_default:
          type: boolean
    DailyActionPlan:
      type: object
      properties:
        day:
          type: integer
        actions:
          type: array
          items:
            $ref: '#/components/schemas/WarmingAction'
    WarmingAction:
      type: object
      properties:
        type:
          type: string
          enum:
            - view_profile
            - like_post
            - subscribe
            - comment
            - send_message
            - repost
            - view_stories
            - add_friend
        count:
          type: integer
          minimum: 1
        weight:
          type: number
          format: float
          description: Вероятность выполнения действия
    StartWarmingRequest:
      type: object
      required:
        - account_id
      properties:
        account_id:
          type: string
        scenario_id:
          type: string
          description: ID сценария (по умолчанию - basic)
        duration_days:
          type: integer
          minimum: 1
          maximum: 30
          description: Переопределение длительности сценария
    CreateScenarioRequest:
      type: object
      required:
        - name
        - platform
        - daily_actions
      properties:
        name:
          type: string
          minLength: 3
          maxLength: 100
        description:
          type: string
          maxLength: 500
        platform:
          $ref: '#/components/schemas/Platform'
        daily_actions:
          type: array
          items:
            $ref: '#/components/schemas/DailyActionPlan'

    # ==================== SMS Schemas ====================
...